pyproject.toml
[project]
name = "semantic-vision"
version = "1.0.0"
description = "AI-powered visual testing framework using Google Gemini and Playwright"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["testing", "visual-testing", "ai", "playwright", "gemini", "qa", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Testing",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "google-generativeai>=0.3.0",
    "pytest>=7.4.0",
    "pytest-html>=4.0.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/semantic-vision"
Documentation = "https://github.com/yourusername/semantic-vision#readme"
Repository = "https://github.com/yourusername/semantic-vision"
Issues = "https://github.com/yourusername/semantic-vision/issues"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-s --html=reports/report.html --self-contained-html"
log_cli = true
log_cli_level = "INFO"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.isort]
profile = "black"
line_length = 100
